Use size_t to report file size and position
This way it is consistent with the GSM_File structure.

+++ include/gammu-file.h
@@ -224,7 +224,7 @@ GSM_Error GSM_SetFileAttributes(GSM_StateMachine * s, GSM_File * File);
  * \ingroup File
  */
 GSM_Error GSM_GetFilePart(GSM_StateMachine * s, GSM_File * File, int *Handle,
-			  int *Size);
+			  size_t *Size);
 
 /**
  * Adds file to filesystem. Call repeatedly until function returns
@@ -240,7 +240,7 @@ GSM_Error GSM_GetFilePart(GSM_StateMachine * s, GSM_File * File, int *Handle,
  *
  * \ingroup File
  */
-GSM_Error GSM_AddFilePart(GSM_StateMachine * s, GSM_File * File, int *Pos,
+GSM_Error GSM_AddFilePart(GSM_StateMachine * s, GSM_File * File, size_t *Pos,
 			  int *Handle);
 /**
  * Sends file to phone, it's up to phone to decide what to do with it.
@@ -257,7 +257,7 @@ GSM_Error GSM_AddFilePart(GSM_StateMachine * s, GSM_File * File, int *Pos,
  *
  * \ingroup File
  */
-GSM_Error GSM_SendFilePart(GSM_StateMachine * s, GSM_File * File, int *Pos,
+GSM_Error GSM_SendFilePart(GSM_StateMachine * s, GSM_File * File, size_t *Pos,
 			   int *Handle);
 /**
  * Acquires filesystem status.
