+++ libgammu/api.c
@@ -1742,7 +1742,7 @@ GSM_Error GSM_SetFileAttributes(GSM_StateMachine *s, GSM_File *File)
 /**
  * Retrieves file part.
  */
-GSM_Error GSM_GetFilePart(GSM_StateMachine *s, GSM_File *File, int *Handle, int *Size)
+GSM_Error GSM_GetFilePart(GSM_StateMachine *s, GSM_File *File, int *Handle, size_t *Size)
 {
 	GSM_Error err;
 
@@ -1755,7 +1755,7 @@ GSM_Error GSM_GetFilePart(GSM_StateMachine *s, GSM_File *File, int *Handle, int 
 /**
  * Adds file part to filesystem.
  */
-GSM_Error GSM_AddFilePart(GSM_StateMachine *s, GSM_File *File, int *Pos, int *Handle)
+GSM_Error GSM_AddFilePart(GSM_StateMachine *s, GSM_File *File, size_t *Pos, int *Handle)
 {
 	GSM_Error err;
 
@@ -1768,7 +1768,7 @@ GSM_Error GSM_AddFilePart(GSM_StateMachine *s, GSM_File *File, int *Pos, int *Ha
 /**
  * Sends file to phone, it's up to phone to decide what to do with it.
  */
-GSM_Error GSM_SendFilePart(GSM_StateMachine *s, GSM_File *File, int *Pos, int *Handle)
+GSM_Error GSM_SendFilePart(GSM_StateMachine *s, GSM_File *File, size_t *Pos, int *Handle)
 {
 	GSM_Error err;
